one class per header file
